Add infix functions, and the require keyword.
More specifically, the changes in this commit are the following:

* add infix functions on predicates, accessible in every module;
* generate an equality function for every new declared type;
* implement the 'require' keyword and the ability to reduce a module to
  its signature;
* fix some minor bugs in the compiler.

Note that the infix functions on the predicates are now hardcoded in
the compiler. It is likely that this is not how we want to implement
this in the long run. However, since this is somewhat of a blocker and
since that part of the design has not been finalized, we implement it
like so for the moment.