Implement fully qualified names handling in the compiler.
References to top-level elements are now resolved using the following
rules:
(1) if the dependency name is fully qualified, resolution succeeds if an
    element with the same fully qualified name is in scope;
(2) if the dependency name is not fully qualified (i.e. a prefix
    package name is not provided), then resolution succeeds if
    (a) there is exactly one top-level element with the same
        (non-fully-qualified) name in scope;
    (b) there is an element with the same (non-fully-qualified) name
        defined in the local scope. In this case, even if other elements
        share the same name and are in scope, the locally defined one
        takes precedence.

An additional note for (2.b): as it is now, there is no way to fully
qualify local definitions. This is because the package name is not 'in
scope', so to say.