Implement facilities for error recovery and MgMonad.
Errors thrown during compilation are now aggregated within a State
contained in the newly introduced Magnolia monad. The intended use is
for the topmost MgMonad to fail if any subnode has produced an error,
while still traversing each branch of the AST as deep as possible (i.e.
until the current phase succeeds at typechecking the whole subtree, or
until an error is thrown somewhere in the subtree).