Flesh out required and concrete types handling.
The implementation given here is rather brittle, and should be improved
once a proper core language for the Magnolia compiler is created.

Handling required types is blocking for code generation however, hence
why this initial implementation is provided.

This commit also fixes the source location for errors related to 'uses'.