Improve the internal handling of requirements and implementations.
The distinction between concrete declarations (e.g. external types,
callables with bodies) and abstract declarations (e.g. required types,
function prototypes) was not clearly enforced within the compiler;
before, we allowed our scope to contain several declarations for a given
callable, namely one for the prototype, and one for the implementation.
This required some careful manipulations of the scope when registering
new types/prototypes, which was brittle and undesirable.

This commit modifies the type of annotations of module-level
declarations such that they now carry around a (non-empty) list of
abstract declaration origins, as well as maybe one concrete declaration
origin, such that:
    * each concrete declaration produces both a concrete and an abstract
      declaration origin;
    * each abstract declaration produces only an abstract declaration
      origin;
    * two declarations can be merged if at least one of them is abstract
      OR if they correspond to the same concrete declaration (e.g. if
      a function is imported through different paths with the same name,
      for instance).

With this, we can now carry around one unique node in scope
corresponding to both the abstract & concrete version of a given
declaration. Any duplicate should now be a compiler error.

We perform some hacks to deal with the "hackyPrelude", which should not
be looked too hard into.