Improve name resolution for dependencies at both the package and module level.
With this setup, module and package dependencies are not statically
guaranteed to be populated with their "qualifying" information. One
solution for that would be to add yet another type family. However,
since these dependencies are updated in very specific places, it seemed
to make sense to avoid the complication brought by the type family and
to instead add some safeguards throughout the code. This decision might
be revisited later.

Fixes #7.