Separate compiler passes more explicitly.
To be able to do that, PackageHeads also now carry source information
that can be queried. The compiler can now be called using the "test"
subcommand to test compilation up to a specific compiler pass, and
for a specific backend (once codegen will be implemented).

The commit also cleans up the legacy WithSrc-related utils, which are no
longer needed. Various additional cleanups are forthcoming, but not
included here, so as to keep the resulting PR readable.