Implement a first version of (non structure-preserving) code generation for a C++ backend for Magnolia.
Functionality is for the moment limited. Code is only generated for
programs, and we assume that externals do not have template parameters.
This latter point will be taken care in a forthcoming PR.

It is difficult to produce good and extensive tests for code generation.
This will be tested more thoroughly as we write example code in Magnolia
and document the failures we encounter.

Forthcoming PRs should also add some codegen design documentation to the
repo.