Implements partial and total renamings.
Two types of renamings can now be applied. Assuming a module A, one can
write `A[Z => Z]` or `A[[Z => Z]]` to apply respectively a total and a
partial renaming block to A.

The expectations are that total renaming blocks expect source names in
the block to correspond to the name of at least one declaration in `A`,
while partial renaming blocks have no such expectations, and therefore
allow renamings with non-existent sources.