Handle module casts, and handle "built-ins" functions a little bit better.
To deliver signature casting, of module expressions, care must be taken
when it comees to the handling of "built-in" operations (i.e. predicate
combinators, equality operator generated for types, etc).

The solution essentially relies on an additional constructor for
"ConcreteDeclOrigin"s, and an additional constructor for "CBody"s. One
unsatisfying aspect of the solution is that a callable annotated with a
concrete "GeneratedBuiltin" declaration can be used as both a prototype
and an implemented function, and it is no longer sufficient to check if
"isNothing $ fst $ _ann callable" to know if a callable is implemented.

This abstraction is still not perfect, but should do for the time being,
since delivering signature casting has become critical. TODOs are left
to revisit the issue at a later time.