Clean up the implementation of External modules.
Externals are now handled in a more principled fashion, as a sort of
'module expression functor'. This ensures that we can now write things
like 'program P2 = external C++ path.to.file.struct signature(P1)',
which could not be handled using the previous definition of externals.

We also strengthen the external annotations by explicitly constructing
instances of a the new 'ExternalDeclDetails' data type. This reduces the
need to write partial code handling FullyQualifiedNames, which are not
statically required to contain a scope name.

The implementation of this latter feature required a bit of
reorganization of the repository, to resolve cyclic dependencies.