Implement handling of function requirements in C++ code generation.
An additional example is added to the repository, corresponding to a
while_loop implemented in Magnolia for two different iteration types
(Int16 = short, Int32 = int).

We also fix a bug in the type checking phase that allowed merging
prototypes and bodies differently, sometimes producing callables with
different argument names than those used in the body of the callable.

One remaining caveat with this implementation of function requirements
is that calling a function overloaded solely on return type now requires
calling f.operator()<T>(…) explicitly, instead of the simpler f<T>().
There is a plan in place to fix that in the future.