Sort typedefs topologically and add forward declarations of function call operator-implementing structures.
This replaces the dummy data structures we previously implemented, which
were actually not working as intended; the reason for that is that
Magnolia only defines opaque types, which may depend on the type
requirements of the external module considered. We now assume that all
these types can be defined in dependency order, and throw an error if a
strongly connected component exists.